Set a timeout for git to prevent tags being set
Yes, that's right. With no timeout, the <extensions> tag is left out of
the job config. With no <extensions> tag, the GitSCM plugin assumes
defaults for it which includes setting a tag per build. Setting a tag
per build breaks the "python setup.py --version" because it uses tag
data.